pyproject.toml
[project]
name = "hogart-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "telethon>=1.36.0",
    "load-dotenv>=0.1.0",
    "langgraph>=0.2.22",
    "langchain-experimental>=0.3.0",
    "langsmith>=0.1.120",
    "pandas>=2.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
